Refactor: streaming frequency 타입 변경

+++ capstone1/src/main/java/org/vox/capstonedesign1/dto/AgentViewResponse.java
@@ -13,7 +13,7 @@ public class AgentViewResponse {
     private final String estimatedStatus;
     private final String deviceSerialNumber;
     private final LocalDateTime timeStamp;
-    private final int streamingFrequency;
+    private final double streamingFrequency;
 
     public AgentViewResponse(Agent agent, AgentSignal agentSignal) {
         this.agentName = agent.getAgentName();
